# server-go/doc/swagger/swagger.yaml
basePath: /v1
definitions:
...
  controller.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  controller.GetProfileDataResponse:
    properties:
      avatarUrl:
        type: string
      email:
        type: string
      fullname:
        type: string
      isUserProfile:
        type: boolean
      nickname:
        type: string
      numFollowers:
        type: integer
      numFollowing:
        type: integer
      userID:
        type: integer
    type: object
...
  controller.SuccessResponse:
    properties:
      message:
        type: string
    type: object
  controller.UpdateProfileResponse:
    properties:
      avatarUrl:
        type: string
      email:
        type: string
      fullname:
        type: string
      nickname:
        type: string
      userID:
        type: integer
    type: object
...
host: localhost:8000
info:
  contact: {}
  description: server-go
  title: Instagram-lite-copy API
  version: "1.0"
paths:
...
  /profiles/find/{name}:
    get:
      consumes:
      - application/json
      description: Find users by nickname with pagination
      parameters:
      - description: User nickname
        in: path
        name: name
        required: true
        type: string
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: Offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              properties:
                avatarUrl:
                  type: string
                fullName:
                  type: string
                nickName:
                  type: string
                userId:
                  type: integer
              type: object
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Find users by nickname
      tags:
      - Profile
  /profiles/id/{id}:
    get:
      consumes:
      - application/json
      description: Get profile data for the user with the given ID or nickname or
        nothing if user is autherization
      parameters:
      - description: User nickname
        in: path
        name: nickname
        type: string
      - description: User ID
        in: path
        name: id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.GetProfileDataResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get profile data
      tags:
      - Profile
  /profiles/id/{id}/follow:
    put:
      consumes:
      - application/json
      description: Toggle follow/unfollow user by user ID
      parameters:
      - description: User ID to follow/unfollow
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Toggle follow/unfollow user
      tags:
      - Profile
  /profiles/me:
    get:
      consumes:
      - application/json
      description: Get profile data for the user with the given ID or nickname or
        nothing if user is autherization
      parameters:
      - description: User nickname
        in: path
        name: nickname
        type: string
      - description: User ID
        in: path
        name: id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.GetProfileDataResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get profile data
      tags:
      - Profile
    patch:
      consumes:
      - multipart/form-data
      description: Update the profile information for the authenticated user
      parameters:
      - description: User full name
        in: formData
        name: fullname
        type: string
      - description: User email address
        in: formData
        name: email
        type: string
      - description: User nickname
        in: formData
        name: nickname
        type: string
      - description: User avatar image
        in: formData
        name: avatarImage
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.UpdateProfileResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Update user profile
      tags:
      - Profile
  /profiles/nickname/{nickname}:
    get:
      consumes:
      - application/json
      description: Get profile data for the user with the given ID or nickname or
        nothing if user is autherization
      parameters:
      - description: User nickname
        in: path
        name: nickname
        type: string
      - description: User ID
        in: path
        name: id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.GetProfileDataResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get profile data
      tags:
      - Profile
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"